pref: used signals from preact
- used signals from preact to avoid the need for separating ClockHands
into a separate component, and using signals only hands are rerended
