+++ package.json
@@ -31,6 +31,7 @@
     "vite-tsconfig-paths": "3.5.0"
   },
   "dependencies": {
+    "@preact/signals-react": "^1.0.2",
     "react": "18.2.0",
     "react-dom": "18.2.0"
   },
